Add Assignment class and list of Assignments to Student class
This commit adds a new Assignment class to the student package, which encapsulates the description and deadline of a particular assignment. The class is implemented with a constructor that takes the required fields as parameters, as well as getters to access the fields.

In addition, this commit modifies the Student class to include a list of Assignment objects. This list is initialized in the constructor with an empty ArrayList, and there are corresponding methods for adding and removing assignments from the list. The toString method of the Student class has also been updated to include the list of assignments.