Add view-homework command
This commit adds a new command view-homework to the system that allows tutors to view homework for specified students with an option to filter by status. The command uses a ViewHomeworkCommand class to handle the logic of filtering homework and a ViewHomeworkCommandParser class to parse user input.

The command takes in optional parameters n/STUDENT_NAME and st/STATUS that allow tutors to filter homework by student name and status respectively. By default, all homework will be displayed. The available status values are completed and pending.