Add Storage of Assignments
This commit updates the storage function in TutorPro to allow for the storage of assignments for each student. The Student class has been updated to include a List of Homework objects, representing the assignments for that student. The StorageManager class has been updated to support the saving and loading of these Homework objects in the data file. Additionally, the ModelManager class has been updated to provide methods for adding and removing Homework objects from a Student's list of assignments, and for retrieving the list of assignments for a given Student. These changes enable TutorPro to more effectively manage student assignments and progress.