Add method to StringUtil to accomodate substring searches for "find"
The original find command uses containsWordIgnoreCase, which only
accepts full-word searches.

However, when we want to search names, we may not recall the full name.
For example, we may remember Bernice as "Bern..."

Therefore, as a step to more efficient searches, we will check if the
search is a substring of the full name first.