add functionality to search by postal code with starting numbers
Instead of just being able to search by name, the user should be able to
search for dengue cases by postal codes. In this case, it would be good
to search for postal codes which start with a certain index.
Additionally, it is essential to consider postal codes which do not
simply contain a certain sequence of numbers, as this may cause too few
results to be filtered