Add noisy JSON reading method
Logger is being used to convey the message of what went wrong while
loading a JSON file.

Logger messages are difficult to pass around to other components.

Lets create a method that reads a JSON file but throws an exception when
it encounters a problem. This way the exception can be caught and the
message can be displayed to relevant components of the application.