fix: Fix blurry contents in scrollpane
Adding scrollpanes to view causes its contents to be blurred. This is due
to `ScrollPaneSkin` setting cache to true on intialisation, hence
ignoring the cache property set in the FXML file.

This fix aggressively attempts to reset this property in the
scrollpane. However, the skin in the scrollpane must first be initialised, hence the need to
repeatedly attempt to check for its existence.