Update sorts Update sorts to handle empty strings.
Currently, the sort makes use of natural ordering which places empty string at the front of the list. This is counter intuitive to the purpose of organising the list based on a category.

The user would want contacts that does not have the specified category to be at the back.

Let's
* update sort in UniquePersonList to place empty strings last.

UniquePersonList.java:
* update comparators to position empty strings at the end.