Add function to hide label when not in use
Current calorie label will create an empty space.

Removing the empty space when it is not in used will be good
for the UI.

Let's create a condition for the calorie display and
use setManaged method to determine if it should be displayed.

setManaged method allows the fxml file to stop managing
the label which is the only way to hide the label if needed.