Rename package from seedu.address to vimification
The old package name is seedu.address.

We need to change the name to make it clear that this is the code base for Vimification.
Also, changing the package name removes 1 level of directories. The extra level of nested
directory is not necessary.

The package is renamed from seedu.address to vimification. Occurrences of seedu.address in all
files are also replaced with vimification.

This is a big change that affects the entire codebase. Testing is required to ensure that the
code still works correctly after the renaming.