Add Routine and Exercise class with their storage system.
Current program has no Routine and Exercise class.

Required to create storage system for Routine and Exercise class as it is
a different entity.

Let's add storage for Exercise and Routine. Additional Commands for Routine:
AddRoutineCommand and EditRoutineCommand added. Storage with sample data included.

Adding a storage system is better for Exercise and Routine to not mix it with
the clients storage. Routine and exercise should be its own entity. However, for
future uses, there should be a link between them.

Current Todo/Issue:
- Testings need to be included.
- Delete Routine, Find Routine
- Marking / Unmarking of exercises in routine when tagged to clients.