Add find command to include finding Address of a person. Add test cases to test for this new functionality.
Needs to be implemented so that we may search for health conditions/medicine usage in future.

Needs to include "a"/"n" to indicate address attribute or name attribute before the actual arguments. Eg. "find a Bernice".

It is done this way to differentiate the different attributes that we might search for in future.