Update FindCommandTest and AddressBookParserTest
Use of containsWordIgnoreCase requires the word parameter to not be
empty string which was what was being used in AddressBookParserTest test
case thus causing it to fail. Directly enter lists into
InternshipContainsKeywordPredicate instead of entering strings into
private method that in turn create InternshipContainsKeywordPredicate
object. This will allow for empty lists to be used instead of empty
string being passed into private method which then creates a list of
size 1 with element empty string.

parseCommand_find in AddressBookParserTest was not adhering to the
status and tag input formats of single alphanumeric word.