Add tag adding and editing functionality
Users are unable to tag a dengue case with the associated dengue variant.

Adding dengue variant tags will allow for more granular analysis of the commonness of different dengue strains in Singapore, and enable quick action should a particular strain that residents are less immune to become unusually common.

Let’s add the ability to add and edit tags that will indicate the dengue variant associated with each dengue case.

Using an enum will allow us to easily validate that only actual dengue variants are used as tags compared to using regex or a list of accepted input.