Change editRoutine command
Current editRoutine command functions like edit command.

Required to change so that user can select which routine to change the
name or select which routine and the exercise number to change which
exercise within the routine. This is important because we do not want
the users to retype every exercises to change one exercise or routine
name.

Let's change the Set for exercises to an ArrayList and change up the
editCommand format to handle any errors and also compare the
editedDescriptor with the current routine selected.

Usage of Arraylist is better than Set as Set is unordered and this
program requires an ordered list to do adding and deleting of certain
exercises. The rationale behind this is to prevent users from clearing
the entire list of exercises in a routine when making edits.

Current Issues to be handled (Coming soon):
- Testings