Add editing of status to Person class and basic UI representation.
For each contact, we would like to assign common statuses to help
the user track how far into the sales funnel the person has reached.
4 types of lead statuses have been identified.

Let's restrict the user's ability to choose the lead status of each
contact to four choices, to ensure that the descriptors are consistent
and categorisable.