Add 'meetingAdd' command
Implement meetingAdd command to add meetings to clients in FAid.

Update relevant classes to adopt the use of Meeting class.

UI for meeting has been temporarily left out as awaiting for development of UI from the other side.

Update meeting class to be used for testing. Update to test cases to be done in a later commit.