Add support for Transactions in Model and AddressBook.
Currently, AddressBook stores data for Persons only. To manage
Transactions and the relationships between Persons and Transactions,
the data for Transactions and relations should be stored in
AddressBook as well.

Let's add a list of transactions (UniqueTransactionList) and relations
(PersonTransactionMap) to AddressBook attributes, and the skeletal
methods to support the operations of these attributes.

Let's also add abstract methods to the Model to support the operations
on these attributes, and skeletal code to the ModelManager to facilitate
the implementations of these operations.