Check function - version 2
Listview scrollTo is inaccurate, change the implementation approach
The user can check for one student's information, meanwhile, instead
of display the entire list, will only display the checked student.

Modified to UI to resolve problem, when the number of cell
is not enough, the background will auto filled with the empty list
view with background color. Change them to transparent.