DeleteCommand class deletes by Index in the list
While Index works fine, it is tedious to look for the particular person index if the list is very long

Let's change the delete command to delete by name. If there are multiple people sharing a common name, the gui prompts the user to enter the specific name to delete.

The user no longer have to search for the person's index