Fix bug - all match
The previous check is based on the existing NameContainsKeywordsPredicate,
however, the behind logic is to use any match, which defeat the purpose of check
as the intended behaviour should be all match.

Therefore, the code is now changed to all match (as we will not have people with exact
same name, only one person can be selected each time).