Allow user to use keywords to filter out irrelevant contacts
As the user may want to find contacts with fields containing certain
keywords, let's add a new function that enables filtering. Let's add a
new predicate that allows us to search for substrings in a field that
enables this new feature.

Let's also add tests to test this feature out.