Add support for Undo Command in ModCheck
Modcheck does not support undo.

User who accidentally made typos has to reverse their changes manually.
This may not be easy to do in some cases (Eg: clear).

Lets add a UndoManager into ModelManager.
That acts as a simple version control to support undo and redo.

In the future, support for redo can be added.
Optimisations can also be done to reduce memory usage of UndoManager.