Add ViewMode and ListMode
Due to the desire to have a user story with a GUI detailed view of each entity sheet, a mode system is required to facilitate the switching between the standard list mode and view mode displays in the GUI.

Several UI elements were added in the FXML template for MainWindow to assist in the arrangement of these new UI modes.

The previous UI was moved to ListModelPanel and a new ViewModePanel was created with additional space for detailed entity information display.

A new "view" command was added to toggle between the 2 display modes.

#22