Refactor test cases to use ModelManagerBuilder.
Currently, test cases uses a different version of constructor of ModelManager that takes in FriendlyLink instead of Storage.

This requires the use of an additonal constructor in ModelManager that is not used in production.

Let's refactor this constructor in to a test utility class ModelManagerBuilder to create ModelMangers using FriendlyLink instead of Storage.