Remove lead status search from StatusCommmand.
Previously, the status command was to support searching by status
categories. However, this would be the first deviation from the 'find'
command, and filtering would be done in a within the 'status' command.

Let's remove the searching from StatusCommand, and leave the
implementation of search by lead status to the FindCommand.
StatusCommand is now strictly for the setting of lead statuses of
contacts.