Rename AddressBook to TeamBulder for all classes
All classes are currently named related to address book.

We are now making this project called Team Builder and the relation to
address book is no longer sensible.

Let's rename all the classes to team builder and also change the project
file structure to better reflect that this is no longer a seedu.address
project.

This will allow us to better communicate as class names and structure is
now more sensible.