Update Help Window
Current help window is showing AB3 User Guide, links to AB3 User Guide

This is not the User Guide for FastTrack and does not show the correct
commands for using FastTrack

Let's change the URL to link to FastTrack's (temporary) User Guide, and
also add a simple list of commands in the Help Window.

This allows users to get a quick overview of the commands they can use,
and have a correct link to the right user guide if they want to know
more.