Session - Json Format & UI
To use the Session class, simply create a new instance with the appropriate start and end times. Example inputs are strings in the format "dd-MM-yyyy HH:mm". The class can then be used to compute the duration of the session.

The JSON format for storing Person objects has been modified to include a session field, which contains information about the session in the format "dd-MM-yyyy HH:mm to dd-MM-yyyy HH:mm".

Example: add n/John Doe p/98765432 a/311, Clementi Ave 2, #02-25 r/35 s/10-03-2022 10:00 to 10-03-2022 11:00 t/friends t/owesMoney

s/10-03-2022 10:00 to 10-03-2022 11:00

The UI has been updated to display session information for each Person in the main view. The PersonCard now includes a label for session start and end times, as well as the duration of the session.