Refactor Undo and Redo to use Undoable interface.
Undo and Redo calls a few extra functions in ModelManager.
These functions are not originally part of the Model interface and had to be added.

The Model interface suffers from high coupling. Implementing the Model interface
 necessitates implementing the undo functionality, which may not be desirable.

Let's extract the undo functionality into an Undoable interface.
The ModelManager will implement both the Model and Undoable interface.

This allows future models to be developed without the need for undo functionality.
It also allows other objects to be marked undoable.