Migrate Person to Entity
To make the context of the classes fit better with the TTRPG purpose
this project serves, the original "Person" class and all methods,
variables and tests need to be altered to fit the new core "Entity"
class.

Let's refactor most of the main classes first to fit the new logic.