Fixed personList disappearing from main stage and errors in testing
Added ViewCommand to show details of selected person instead of modifying FindCommand. Rolled back FindCommand to original one.

Modified DetailsPopup to leave main stage as it was when pop up window is shown so that the user can still use the main stage after he closes the pop up stage.