Implement skeleton for edit appointment command
Add parser and placeholder logic for the edit appointment command. Placeholder logic uses index of a patient in the list instead of the actual patient, and the parser parses the index instead of ID. It requires the model to be passed from LogicManager to AddressBookParser for now. This will be removed once patient ID is implemented.