Update CI to not run on draft PRs
Do see the following links on where the solution was gathered from.

This change allows the PR to stop having red X everywhere when it is not
ready for review yet. PR drafts allow for team members to catch up on
what is being worked on. So that we can be updated on what is happening.

https://gist.github.com/myobie/671d2f2a1e503efa0c05e7b865c26a9d

https://github.com/orgs/community/discussions/25722#discussioncomment-3248917