Fix Gradle build failures
There're issues with loading the Patient and Doctor saved details in the
address book due to the use of inheritance of Person. Currently cannot
identify the root cause of the bug.

Quick fix is to include the patient fields (ie. List<Appointments>) in
the Person class to bypass the use of JsonAdaptedPatient,
JsonAdaptedDoctor which is causing the storage problems.

Let's add appointments as an additional field to Person.