Implement list and edit appointment commands
Add `list_appt` command to list all appointments. Since appointments are not implemented in the GUI yet, the list of appointments will be output in the terminal instead.

Add `edit_appt` command to edit an appointment. Appointment IDs are shared with patient IDs so the first appointment ID will be the lowest unused integer after all the patient IDs have been initialised. In the sample data, this will be 7. References to the patient using patient ID has not been fully implemented yet, so the patient ID might change to an arbitrary number for now.