Change how start time within days is determined
Determining if a start time is within the given number of days is done
by calculating the full days between the times.

Changing these to considering dates but not hours and minutes is more
intuitive for the user, as they will likely only consider the date when
using this command.

Let's adjust the method so that if the start date is earlier than
or equal to the date the given number of days from now, it returns true.
It still returns false if the event starts earlier than now, considering
the full time given.