Add tests for EventBookParser.java
It is difficult to check if two RemindCommand are equal if one is
created with the time now. So it is difficult to test a RemindCommand
created by EventBookParser, which uses time now.

Passing the time into the EventBookParser will make it easier to test if
it is parsing a RemindCommand correctly.

Let's create a constructor for EventBookParser that takes a
LocalDateTime timeNow as an argument, and test EventBookParser by
passing in a typical time in place of the actual time now.