DeleteCommand class deletes by name in the list
While name works fine, it is not secure and does not handle the case of having common shared names. If the list contains "John" and "John Doe" then delete n/John will delete "John". However the user might want to delete "John Doe" but accidentally pressed enter before typing out the rest of the name and the wrong "John" will be deleted.

Let's change the delete command to delete by NRIC. Since each person has a unique NRIC, there is no need to check for multiple person. The GUI will automatically delete the person with the respective NRIC.

The user no longer has to search for the person's index or name.