Update HistoryUtil to better apply Demeter's Law
Concrete Memento has a string description that the UndoCommand uses to
provide feedback to user. However, a memento cannot have a string description
when made as it is simply a state of the originator and has no knowledge
of time and change. Hence, the memento's description setter is used by
the modifying Commands to descibe the change occured.

This is problematic as this violates Demeter's Law. The Commands use a
method from the memento from the method from model to set a description.
Furthermore, this complicates the Commands' relationship with Memento as
Commands now directly rely on the Concrete implementations of Memento to
set descriptions.

Let's move the descriptions out of the memento and into HistoryUtil. We
will now store the description of the change, along with the memento, into
HistoryUtil instead. HistoryUtil should directly interface with memento,
while Commands should directly interface with model and HistoryUtil instead.

This will resolve Demeter's Law as a memento's methods are no longer used
by the Commands. Command and Memento classes are also decoupled for better
extension in accordance to the open-close principle.

Misc.
- Changed Momento to Memento (spelling).
- Changed HistoryUtil.store to HistoryUtil.storePast (better name).
- Updated Tests to reflect above changes.
- Updated Commands to reflect above changes.