Update link from routine to client
Current program has no link of client with routine.

Linking of client to certain routines help to fix some issues.

Let's edit the Add Command to add routines to client and modify json
files to suit this procedure.

This approach is better to decide which client require which routine
exercise.