Update more functionality for commands to suit new Routine to Client link
Current program cannot adapt to new linkage between Routine and Client.

Required to update these commands to work well with Routine and Client.

Let's change the commands to work with the model to make any changes
if required for the clients' routines.

Better to do most of these features in execute to allow modifications in
model.