Add sort Event feature
Events are display in the order in which they were added.

User might want to see the new few upcoming events which they may
need to pay more attention to.

Let's allow Events to be compare to each other (refer to previous
commit in this branch), and add a Sort command.

Another implementation not adopted was to actively add Event in
order. This was not chosen since it might introduce lag to the add
process when the list is lengthy.

This commit will also make the feature to display the next event to
the user easier. Tests for this feature will be introduced in a
later commit.