Fix bug causing 'AddressBookParserTest::parseCommand_edit' test to fail
Test command had the arguments "s/t/" which was being parsed as a
subject arg. with the value "t/", instead of an empty subject and tag
arg.

Caused by AB3 team not suffixing their "t/" arg with space, and when
that tag code was copied by us, the "s/" arg also didn't have a space.