Notification of reminders can show at every clock hour.
Notifications of reminders only showed when the app start up. Having the
notification show the reminders at every hour helps with informing the
user better.

Let's:
* Utilise TimerTask and Timer to schedule a check in the reminder list
at every clock hour

I use TimerTask and Timer because running a Thread that constantly
checks the current time blocks the app and causes it to hang.