Edit the date class to accept new formats for dates Dates should adhere to the following constraints:
Dates to contain 4-digit years YYYY, 2-digit days and either 2-digit
months MM, 3-letter months MMM, or fully-spelled months MMMM.
Special characters may now be allowed (no longer dashes to separate YYYY
from MM), and the ordering of data will not matter unless the data is
ambiguous. (See below)
If months are given in the MM format and the dates are ambiguous, eg.
20001010, YYYY will be read first, followed by MM and DD.
If months are given in MMM or MMMM formats, months will be read first,
eg. march 20 10 february will be read as march
Dates must be valid. i.e. MM must be a number between 01 and 12
inclusive, dd must be a number between 01 and 31 inclusive
dd should be within the valid range for the MM given