Resolve issue with wrong indexing after sorting
Currently, there are issues with applications in the GUI having
a misaligned index with the internal indexing of the system.
For example, if a user requests to edit application at index 3,
an application at a different index on the GUI will appear to be edited.

Let's,
- Modify existing commands to update with sorted list
- Add a DefaultComparator that sorts lists based on their order of
creation
- Modify list command so that it shows a "sorted" list that is
"sorted" with the DefaultComparator