Remove case "n/" from ArgumentMultimap.java as it should not be optional and it is not useful for StudentEditCommand
Fix StudentAddCommand.java display message spacing and minor error
in setting the Parent for the Student

Fix StudentCommandParser.java by removing name from
StudentEditCommand() as it is not used at all

Partial FIX of StudentEditCommand.java

Notes:
- There is some error in the logic of StudentEditCommand
- When editing Student particulars, if Parent Name and Parent Number
not changed, simply update the Student and update the Parent binded
with the Student so that the Parent has updated Student info
- If Parent Name change but Parent Number same, FOR ALL STUDENTS
under the parent, need to update the parent given for the students
with the new updated parent name.
- If different parent number, remove binding of the Student from the
parent and bind the Student to the parent with the NEW PARENT NUMBER
- If parent number not found, CREATE a new parent and perform
binding of parent and student just like how StudentAddCommand does.