Hide Phone label in Ui if phone field is empty
Previously when a person does not have a phone number, the PersonCard still "reserves" the space where the phone is supposed to be, even though no phone number is shown. To fix this, the setVisible method and setManaged methods are called on the phone if the person's phone field is empty.