Fix behaviour for AddCommandParser and EditCommand parser
Allow AddCommandParser to accept inputs with no phone number. This is done by removing `PREFIX_PHONE` as an argument from the `arePrefixesPresent()` method. An additional check then has to be done to check if `PREFIX_PHONE` is found in the input. If it is, we parse the argument with `parsePhone` method from `PaserUtil` class. If not, we set the phone to be the default empty phone from the phone class. The default empty phone is a final static field, and calling the method `getEmptyPhone` will return the same instance of the empty phone.
Additionally, if an empty string or a string containing only white space is passed as an argument to `parsePhone` method, it will also return the default emptyPhone from the phone class.