Change AddCommandParserTest
We now allow for emails to be omitted, as such the email prefix can be missing. This causes the expected error message to be different. Previously a missing email prefix meant that the user did not place the prefix before the email, which was compulsory. Now when there is no email prefix we take it that the user does not intend to give the person they are adding an email address.

Added tests for expected person that has no email or empty email field. The parse should be successful.