Add Sort Command to Address Book Parser
The user is able to use the command word "sort" to execute the sort
feature.

Major bug detected:
After executing the sort command, the addressbook.json data file appears
to be wiped empty. Seems to be an issue with the logic of the
newly implemented sort() method. Requires immediate fixing.