Refactor warnings to be checked in AddPairCommand instead of in FriendlyLink.
Currently, FriendlyLink checks for differing region of elderly and volunteers to indicate a warning should be thrown.

This implementation prevents other types of warning to be indicated since FriendlyLink can only return true or false instead of indicating the warning.

Let's refactor region check to take place in AddPairCommand instead of FriendlyLink so that it can raise different warnings.