Add storage timestamp functionality to lead statuses
Currently, lead statuses do not have a storage format, and all
statuses of contacts are reset to Uncontacted every time the app runs.

Let's add storage for lead statuses to store the newly set status
of each Person. This way, the app will reflect the updated lead status
each time it runs.

Currently, lead statuses only save the status names of the contacts'
lead status. To allow the user to be able to track when they last
updated the status of a contact (GitHub Issues #62), a new timestamp
feature for lead statuses has been implemented, as a java Instant.

Let's have the Instant of a newly set lead status to be set when the
LeadStatus object is instantiated. Previously, the implementation of
StatusCommand prevents the user from executing the command if the lead
status is the same as the current one. This means that the LeadStatus
will never be created. With this new implementation of timestamps, we
can also ensure that a user cannot refresh the timestamp of a lead
status if their status does not change.

However, this implementation does not stop a user from changing the
status to a new status and changing it back, to refresh the timestamp.
Through my research on other CRMs, this seems like an expected
behaviour anyway. Ideally, a log of the changes to statuses done by the
user should be implemented to reflect their actions.