Update {Delete Lecture Command} and {Delete Video Command} parameter format
- {Delete Lecture Command} requires parameters (in order): LectureName, ModuleCode
where LectureName is the name of the TARGET lecture to be deleted
and ModuleCode is the context of the module where the lecture
to be deleted is within
- {Delete Video Command} requires parameters (in order): VideoName, ModuleCode, LectureName
where VideoName is the name of the TARGET video to be deleted
and ModuleCode and LectureName contain the context of the module
and the lecture that the video to be deleted is within

The next extension of the Delete Commands include the acceptance of multiple TARGETs
for {Delete Module Command}, {Delete Lecture Command} and {Delete Video Command}
to be deleted, by comma separation. This is done through the use of variable arguments
in each {Delete Command}. For variable arguments to be used, the argument has to be
the last parameter passed. Thus, the format has to change.

The format of {Delete Module Command}, {Delete Lecture Command} and {Delete Video Command}
is updated to have the targets as their last accepted parameter. Command object executables
to delete multiple target objects are to be added soon, with the parameter format of
context first, followed by the variable argument targets.

{Delete Module Command}, {Delete Lecture Command} and {Delete Video Command} only take in
one target object to delete but their parameter formats are updated to match the commands
that would delete multiple target objects for consistency purposes, so as to not confuse
readers by having different formats for similar tasks.

Command Executables for Deleting Multiple Target Objects (Lecture, Video) are coming
soon... Command to delete multiple modules was already added in previous iteration.