Update full linkage between exercise routines and client
Current program commands not up to date to match the linkage.

These commands will not work if there is no linkage. Better to update them.

Let's update these commands to match with the linkages.

Updating model in execute of each command allows any cross reference of model
client's content with the current routine and update it.