Enhance sort feature to auto-sort when events are changed
Sorting requires user to enter a SortCommand.

This can be a hassle to the user, and may produced unexpected,
inconsistent results. Eg. user thinks that it had been sorted,
when in actual fact it isnt, thus possibly causing the user to
miss an event.

Let's automatically sort the list of events whenever an event is
added, edited or deleted, to produce a consistent ordering of the
list of events. This is done via a listener.

A listener is a clean way to check for list changes, and can cater
to any future commands that also modify the list.

Subsequently, the SortCommand shall be removed.