New Feature: ShowNext Command
ShowNext command shows the next up coming events.
If a number is provided to the command, the next xx number
(as provided) of events will be shown.

ShowNext command uses the keyword "next".

Checkstyles checks are passed locally.

Testing is done only for parsing of the command.
No test cases for the handling of the command (need to consider how
to create the test case when the outcome is time-dependent).