Upgrade the `find` command and add `only` command
Add the functionality to find either the title, description, tags, ingredients, steps, or even the entire recipe. You can only choose one flag at a time, else it'll throw an error.
- Usage: `find r/ omelettes`
- Usage: `find tag/ delicious`

The only command is to find all the recipes in CookHub that you can cook with only the ingredients you have currently.
- Usage: `only eggs salt pepper`