Change requirements for counting duplicate Persons.
Persons were not allowed to have the same names, but allowed to share
identical phone numbers and emails which should not occur if these
fields are correct.

Let's change the way a duplicate person is detected, such that when a
new person is added the system will check and ensure that person's phone
number or email does not already exists.