Refactor edit task command to share an edit task descriptor
Previously, there were 3 edit decriptors: `editTaskDescriptor`, `editTodoDescriptor` and `editEventDescriptor`.

`editTodoDescriptor` and `editEventDescriptor` take in `editTaskDescriptor` as a parameter in their constructors. This leads to the usage of a several getter and setter methods in each descriptor class.
The code duplication made the codebase difficult to maintain as including a new field would require the developer to change all the 3 descriptors.

Let's:
* use a common `editTaskDescriptor` that holds the data required for both `Todo` and `Event` task objects
* delete `editTodoDescriptor` and `editEventDescriptor`