Update UserGuide based on CS2101 standard
Our UserGuide is based on AB-3 and its format with some changes.

This makes our UserGuide unfocused on our target audience, students
looking to make teams for projects. This also makes our UserGuide score
badly for CS2101.

Let's make our UserGuide more narrative and user-friendly. We will make
sections that are more related to the user's usage and omit details that
are not important to the user. We will also add better formatting,
using the 'alert' classes to provide additional information.

This will make our UserGuide more in-line with the requirements for
CS2101 and more readable as well.

P.S.
This update still does not score fully for CS2101 and
more changes are needed to be more in-line with CS2101's requirements.