Make appointment list only show dr nric for patient, patient nric for dr
Appointment list shown in detailed person panel displays both dr's and
patient's nric. Poor UX as it is confusing as to which to refer to.

Patient's appointment list should only display the nric of the dr that
he is scheduled with. Doctor's appointment list should only display the
nric of the patient that he is scheduled with.

Let's improve the UX for appointment list.