Add snooze function for reminder notifications
Users had an issue with notifications constantly popping up.
Original intent was for the user to delete the reminder as a way to
dismiss the notification. However, this was not intuitive enough.
Hence, have added a snooze function that will stop reminder
notifications from appearing, until a new reminder is triggered.